score monitor funciton added! cpu log output location changed -> /data/cpu/cpulog.txt

+++ android/system/extras/cpuusage/scoreCalc.h
@@ -1,6 +1,6 @@
 typedef struct
 {
-	long cpuUsage;
+	float cpuUsage;
 	int threadUsage;
 	float memoryUsage;
 }RESOURCE_USAGE_T;
